feat(dashboard): extract dashboard widgets into plugins
Implements #98.
