+++ src-tauri/src/plugins/dashboard_widget/ai_usage/mod.rs
@@ -1,6 +1,18 @@
-//! AI Usage dashboard-widget backend module — scaffold only.
-//!
-//! Issue #95 lands only the directory + module declaration. Issue #98 will register an `AiUsageBackend` descriptor (with no required commands — the
-//! widget reads the existing session-store) into the host's [`crate::plugins::PluginRegistry`].
-//
-// TODO(#98): register the AI Usage `DashboardWidgetBackend` descriptor.
+//! AI Usage dashboard-widget backend descriptor.
+
+use crate::plugins::dashboard_widget::DashboardWidgetBackend;
+use crate::plugins::Plugin;
+
+pub struct AiUsageBackend;
+
+impl Plugin for AiUsageBackend {
+    fn id(&self) -> &'static str {
+        "ai-usage"
+    }
+
+    fn display_name(&self) -> &'static str {
+        "AI Usage"
+    }
+}
+
+impl DashboardWidgetBackend for AiUsageBackend {}
