+++ src/plugins/dashboard-widget/git-status/index.tsx
@@ -1,4 +1,10 @@
-// Git Status dashboard-widget module — scaffold only. Sub-issue #98 will
-// move the Git Status panel + its polling hook out of
-// `WorktreeDashboard.tsx` into this directory.
-export {};
+import type { DashboardWidgetPlugin } from '@/plugins/registry';
+
+import { GitStatusWidget } from './widget';
+
+export const gitStatusWidget: DashboardWidgetPlugin = {
+  id: 'git-status',
+  displayName: 'Git Status',
+  order: 100,
+  Component: GitStatusWidget,
+};
