+++ src/plugins/dashboard-widget/index.ts
@@ -1,7 +1,3 @@
-// Dashboard-widget plugin TS interface re-exports — scaffold only.
-//
-// Issue #95 lands only the directory layout. Issue #98 will extract the
-// existing Git Status / AI Usage panels from `WorktreeDashboard.tsx`
-// into the per-widget subdirectories.
+// Dashboard-widget plugin TS interface re-exports.
 
 export type { DashboardWidgetPlugin, DashboardWidgetProps } from '../index';
