+++ src/plugins/registry-context.ts
@@ -1,8 +1,9 @@
 // Plugin registry React context (issue #95).
 //
 // Production wraps `<App />` in `<PluginRegistryProvider>` at the top of
-// `App.tsx`. v1 the provider value is an empty registry; sub-issues
-// #96 / #97 / #98 populate it during construction.
+// `App.tsx`. The app passes a pre-built registry from
+// `createBuiltinsRegistry()` so built-in plugins are available without
+// waiting for runtime registration side effects.
 //
 // This file owns only the context constant. The provider component lives in
 // `./registry-provider.tsx` and the `useRegistry` hook in `./use-registry.ts`;
